Move preferences to /lib
Nothing in preferences needs to be in /app, as nothing relies on rails
anymore
