+++ core/app/models/spree/app_configuration.rb
@@ -17,6 +17,7 @@
 #
 require "spree/core/search/base"
 require "spree/core/search/variant"
+require 'spree/preferences/configuration'
 
 module Spree
   class AppConfiguration < Preferences::Configuration
