+++ core/lib/spree/core.rb
@@ -86,3 +86,7 @@ require 'spree/core/stock_configuration'
 require 'spree/permission_sets'
 
 require 'spree/core/price_migrator'
+
+require 'spree/preferences/store'
+require 'spree/preferences/static_model_preferences'
+require 'spree/preferences/scoped_store'
