+++ core/lib/spree/preferences/preferable_class_methods.rb
