+++ core/lib/spree/preferences/scoped_store.rb
