+++ core/lib/spree/preferences/static_model_preferences.rb
