+++ core/lib/spree/preferences/statically_configurable.rb
