+++ core/lib/spree/preferences/store.rb
